refactor commands dao and services

+++ src/main/java/by/ghoncharko/webproject/command/DeleteBankCardCommand.java
@@ -1,8 +1,12 @@
 package by.ghoncharko.webproject.command;
 
 import by.ghoncharko.webproject.controller.RequestFactory;
+import by.ghoncharko.webproject.entity.RolesHolder;
+import by.ghoncharko.webproject.entity.User;
 import by.ghoncharko.webproject.model.service.BankCardService;
-import by.ghoncharko.webproject.model.service.BankCardServiceImpl;
+
+
+import java.util.Optional;
 
 public class DeleteBankCardCommand implements Command {
     private static final String CARD_ID_PARAM_NAME = "cardId";
@@ -13,14 +17,20 @@ public class DeleteBankCardCommand implements Command {
 
     @Override
     public CommandResponse execute(CommandRequest request) {
-
-        final Integer cardId = Integer.valueOf(request.getParameter(CARD_ID_PARAM_NAME));
-        final BankCardService bankCardService = new BankCardServiceImpl();
-        final boolean isDeleted = bankCardService.deleteByCardId(cardId);
-        if (isDeleted) {
-            return requestFactory.createRedirectResponse(PagePath.INDEX_PATH);
+        final Optional<Object> userFromSession = request.retrieveFromSession("user");
+        if(userFromSession.isPresent()){
+            final User user = (User)userFromSession.get();
+            final Integer cardId = Integer.valueOf(request.getParameter(CARD_ID_PARAM_NAME));
+            final BankCardService bankCardService = BankCardService.getInstance();
+            final boolean isDeleted = bankCardService.deleteByCardId(cardId);
+            final boolean userRoleAsClient = user.getRole().equals(RolesHolder.CLIENT);
+            if (isDeleted && userRoleAsClient) {
+                return requestFactory.createRedirectResponse(PagePath.INDEX_PATH);
+            }
+            //error cannot create
+            return requestFactory.createForwardResponse(PagePath.BANK_CARDS_PAGE_PATH);
         }
-        //todo error message
+        //need authorize
         return requestFactory.createForwardResponse(PagePath.BANK_CARDS_PAGE_PATH);
     }
